refactor: extract cabinet builder
